Fix login length
So lets say you register with the username ... which is 3 characters (the minimum). The discriminator adds 5 characters so you have 8 as the minimum. Now lets say that you have the username Xx69420.DrauliscBreaker42069xX which is 30 characters (the maximum). The discriminator adds 5 characters so you have 35 as the minimum to be able to log into your account. Therefore the correct range for login is 8-35 to let all the people who registered with 3-30 to log in.